move PropertyReader to separate file
