+++ src/main/kotlin/r2dbcfun/R2dbcRepo.kt
@@ -1,14 +1,12 @@
 package r2dbcfun
 
 import io.r2dbc.spi.Connection
-import io.r2dbc.spi.Statement
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.single
 import r2dbcfun.internal.IDHandler
 import kotlin.reflect.KClass
 import kotlin.reflect.KProperty1
 import kotlin.reflect.full.declaredMemberProperties
-import kotlin.reflect.full.isSubclassOf
 
 public interface PK {
     public val id: Long
@@ -78,34 +76,6 @@ public class R2dbcRepo<T : Any, PKClass : PK>(
     public suspend fun <V : Any> findBy(property: KProperty1<T, V>, propertyValue: V): Flow<T> =
         finder.findBy(property, propertyValue)
 
-
-}
-
-internal class PropertyReader<T>(private val property: KProperty1<T, *>) {
-    val name = property.name
-    private val kClass = property.returnType.classifier as KClass<*>
-
-    internal fun bindValueOrNull(
-        statement: Statement,
-        index: Int,
-        instance: T
-    ): Statement {
-        val value = property.call(instance)
-        return try {
-            if (value == null) {
-                val clazz = if (kClass.isSubclassOf(Enum::class)) String::class.java else kClass.java
-                statement.bindNull(index, clazz)
-            } else {
-
-                statement.bind(index, if (value::class.isSubclassOf(Enum::class)) value.toString() else value)
-            }
-        } catch (e: java.lang.IllegalArgumentException) {
-            throw R2dbcRepoException(
-                "error binding value $value to field $name with index $index",
-                e
-            )
-        }
-    }
 
 }
